feat: create mvvm + clean architecture packages and move files
