+++ app/src/main/java/com/picpay/desafio/android/data/User.kt
@@ -1,4 +1,4 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.data
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
