+++ app/src/main/java/com/picpay/desafio/android/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.presentation
 
 import android.Manifest
 import android.content.pm.PackageManager
@@ -12,6 +12,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
+import com.picpay.desafio.android.domain.PicPayService
+import com.picpay.desafio.android.R
+import com.picpay.desafio.android.data.User
 import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Callback
