+++ app/src/main/java/com/picpay/desafio/android/presentation/UserListItemViewHolder.kt
@@ -1,7 +1,9 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.presentation
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
+import com.picpay.desafio.android.R
+import com.picpay.desafio.android.data.User
 import com.picpay.desafio.android.databinding.ListItemUserBinding
 import com.squareup.picasso.Callback
 import com.squareup.picasso.Picasso
